requirements-jetson.txt
absl-py==2.1.0
alembic==1.13.2
astunparse==1.6.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.10
flask==3.0.3
Flask-Migrate==4.0.7
flask-sqlalchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.53.1
future==0.18.2
gast==0.4.0
google-auth==2.32.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
h5py==3.7.0
idna==3.7
importlib-metadata==8.0.0
importlib-resources==6.4.0
itsdangerous==2.2.0
jax==0.4.13
jinja2==3.1.4
joblib==1.4.2
keras==2.12.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
Mako==1.3.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
ml-dtypes==0.2.0
mock==3.0.5
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pandas==2.0.3
pbr==6.0.0
pillow==10.4.0
pkgconfig==1.5.5
protobuf==4.25.3
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1-modules==0.4.0
pybind11==2.13.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.31
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.35.0
termcolor==2.4.0
testresources==2.0.1
threadpoolctl==3.5.0
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
werkzeug==3.0.3
wrapt==1.14.1
zipp==3.19.2
